chore: Move unblock candidate to its own file
